backend/requirements.txt
fastapi
uvicorn
motor
langchain
langchain-google-genai
python-dotenv
pydantic